router improvement, component structure improved, 404 introduced
