+++ client/src/components/LoginForm.vue
@@ -1,7 +1,7 @@
 <template>
   <div class="w-100 h-100 text-center " style="background-color: #333333;">
     <div class="w-50 h-25 m-auto pt-5"/>
-    <div id="login_inputs" class="w-50 h-50 m-auto mt-0 container">
+    <div id="login_inputs" class="w-50 h-50 m-auto mt-0 container" v-on:keyup.enter="login()">
       <router-link
         class="m-0 navbar-brand imi-map-logo"
         :to="{name: 'Home', params: { locale: $route.params.locale }}">
@@ -10,34 +10,35 @@
       <div class="mt-3 text-white">
         Please use your HRZ-Account for logging in.
       </div>
-        <div class="input-group w-auto mt-3 row">
-          <div class="col">
-            <input
-              autofocus="autofocus"
-              class="form-control"
-              placeholder="s0123456@htw-berlin.de"
-              type="email"
-              v-model="username"/>
+        <form v-on:submit.prevent="login()" v-on:keyup.enter="login()">
+          <div class="input-group w-auto mt-3 row">
+            <div class="col">
+              <input
+                autofocus="autofocus"
+                class="form-control"
+                placeholder="s0123456@htw-berlin.de"
+                type="email"
+                v-model="username"/>
+            </div>
           </div>
-        </div>
-        <div class="input-group w-auto mt-1 row">
-          <div class="col">
-            <input
-              autocomplete="off"
-              class="form-control"
-              placeholder="********"
-              type="password"
-              v-model="password"/>
+          <div class="input-group w-auto mt-1 row">
+            <div class="col">
+              <input
+                autocomplete="off"
+                class="form-control"
+                placeholder="********"
+                type="password"
+                v-model="password"/>
+            </div>
           </div>
-        </div>
-        <div id="submit" class="mt-3">
-          <button
-            v-on:keyup.enter="login"
-            v-on:click="login()"
-            class="btn btn-htw-green">
-            Login
-          </button>
-        </div>
+          <div id="submit" class="mt-3">
+            <button
+              v-on:click="login()"
+              class="btn btn-htw-green">
+              Login
+            </button>
+          </div>
+        </form>
     </div>
     <div class="w-75 h-25 m-auto"></div>
   </div>
