+++ client/src/components/internship-module/CreateInternshipModule.vue
@@ -72,7 +72,7 @@ export default defineComponent({
   methods: {
     save() {
       this.$store.dispatch('addNotification', { text: 'Praktikum erfolgreich angelegt!', type: 'success' });
-      this.$router.push({ name: 'InternshipModule' });
+      this.$router.push({ name: 'InternshipModuleIndex' });
     },
   },
 });
