+++ client/src/components/postponements/PostponementsList.vue
@@ -1,0 +1,17 @@
+<template>
+    <div id="form-block4" class="text-left">
+        <h3>Hier folgt die Liste aller Postponements</h3>
+    </div>
+</template>
+
+<script lang="ts">
+import { defineComponent } from 'vue';
+
+export default defineComponent({
+  name: 'PostponementsList',
+});
+</script>
+
+<style lang="scss">
+
+</style>
