+++ client/src/views/Postponements.vue
@@ -1,0 +1,19 @@
+<template>
+  <div class="container clear-top">
+    <div class="margin_bottom_10">
+      <router-view></router-view>
+    </div>
+  </div>
+</template>
+
+<script lang="ts">
+import { defineComponent } from 'vue';
+
+export default defineComponent({
+  name: 'Postponements',
+});
+</script>
+
+<style lang="scss">
+
+</style>
